Lots of changes
Ok well this was supposed to take like 2 hours, it's taken around 4 at this point and still not quite working. There's an issue with the html where it's serving the wrong URL, I can't quite figure out what's going on there. But that's a problem for tomorrow me. For now, goodnight